handler for /upcomingepisodes framework

+++ tvshowdata/tvshowdata.go
@@ -90,7 +90,7 @@ func reformatShowDate(json gjson.Result) (time.Time, error) {
 }
 
 // Determine if there are likely future episodes of a show or not
-func checkForFutureEpisodes(showData string, ID int) (bool, error) {
+func checkForFutureEpisodes(showData string, ID int64) (bool, error) {
 	countdown := gjson.Get(showData, "tvShow.countdown")
 	if !countdown.Exists() {
 		msg := fmt.Sprintf("api returned invalid countdown data for queryID: %d", ID)
@@ -140,7 +140,7 @@ func parseUpcomingEpisodes(showData string) (UpcomingEpisodes, error) {
 }
 
 // Get a list of upcoming shows for a particular Episodate query ID
-func getUpcomingShows(queryID int) (UpcomingEpisodes, error) {
+func getUpcomingShows(queryID int64) (UpcomingEpisodes, error) {
 	url := fmt.Sprintf("https://episodate.com/api/show-details?q=%d", queryID)
 	resp, err := httpGet(url)
 	if err != nil {
@@ -171,7 +171,7 @@ and save everything from then on to add to the calendar
 */
 
 // GetShowData gets the air times of upcoming episodes for the given queryID
-func GetShowData(queryID int) {
+func GetShowData(queryID int64) {
 	episodeList, err := getUpcomingShows(queryID)
 	if err != nil {
 		fmt.Println("Error getting the show data:", err)
